Initialize ata before graphics
ATA init is the long pole in the boot process, and its asynchronous.
move the graphics init after it so that ata and graphics initialize
in parallel

UFS is also a part of ATA and is also somewhat slow in probing.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>
(cherry picked from commit cf141a976bb066f16ca62ee671b4c0ed01773847)
Signed-off-by: dopaemon <polarisdp@gmail.com>